+++ backend/src/main/java/com/example/focusapp/controller/OnboardingController.java
@@ -23,7 +23,7 @@ public class OnboardingController {
     }
 
     @GetMapping("/status")
-    public ResponseEntity<OnboardingStatusResponse> getStatus(@RequestParam Integer userId) {
+    public ResponseEntity<OnboardingStatusResponse> getStatus(@RequestParam Long userId) {
         return ResponseEntity.ok(onboardingService.getStatus(userId));
     }
