+++ backend/src/main/java/com/example/focusapp/dto/CharacterResponse.java
@@ -6,7 +6,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public class CharacterResponse {
-    private Integer characterId;
+    private Long characterId;
     private String imageUrl;
     private Integer level;
 }
