+++ backend/src/main/java/com/example/focusapp/repository/GoalDefinitionRepository.java
@@ -3,7 +3,8 @@ package com.example.focusapp.repository;
 import com.example.focusapp.entity.GoalDefinition;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GoalDefinitionRepository extends JpaRepository<GoalDefinition, Integer> {
-    boolean existsByUserIdAndTitle(Integer userId, String title);
-    boolean existsByUserId(Integer userId);
+public interface GoalDefinitionRepository extends JpaRepository<GoalDefinition, Long> {
+
+    boolean existsByUserIdAndTitle(Long userId, String title);
+    boolean existsByUserId(Long userId);
 }
